+++ jedandroid/src/ch/jeda/platform/android/AndroidCanvasImp.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012 - 2013 by Stefan Rothe
+ * Copyright (C) 2012 - 2014 by Stefan Rothe
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -49,14 +49,14 @@ class AndroidCanvasImp implements CanvasImp {
     }
 
     @Override
-    public void drawCircle(final float x, final float y, final float radius) {
+    public void drawCircle(final int x, final int y, final int radius) {
         assert radius > 0;
 
         this.canvas.drawCircle(x, y, radius, this.strokePaint);
     }
 
     @Override
-    public void drawImage(final float x, final float y, final ImageImp image, final int alpha) {
+    public void drawImage(final int x, final int y, final ImageImp image, final int alpha) {
         assert image != null;
         assert image instanceof AndroidImageImp;
         assert 0 < alpha && alpha <= 255;
@@ -66,12 +66,12 @@ class AndroidCanvasImp implements CanvasImp {
     }
 
     @Override
-    public void drawLine(final float x1, final float y1, final float x2, final float y2) {
+    public void drawLine(final int x1, final int y1, final int x2, final int y2) {
         this.canvas.drawLine(x1, y1, x2, y2, this.strokePaint);
     }
 
     @Override
-    public void drawPolygon(final float[] points) {
+    public void drawPolygon(final int[] points) {
         assert points != null;
         assert points.length >= 6;
         assert points.length % 2 == 0;
@@ -80,12 +80,12 @@ class AndroidCanvasImp implements CanvasImp {
     }
 
     @Override
-    public void drawRectangle(final float x, final float y, final float width, final float height) {
+    public void drawRectangle(final int x, final int y, final int width, final int height) {
         this.canvas.drawRect(x, y, x + width, y + height, this.strokePaint);
     }
 
     @Override
-    public void drawText(final float x, final float y, final String text) {
+    public void drawText(final int x, final int y, final String text) {
         assert text != null;
 
         this.canvas.drawText(text, x, y - (int) this.textPaint.ascent(), this.textPaint);
@@ -97,14 +97,14 @@ class AndroidCanvasImp implements CanvasImp {
     }
 
     @Override
-    public void fillCircle(final float x, final float y, final float radius) {
+    public void fillCircle(final int x, final int y, final int radius) {
         assert radius > 0;
 
         this.canvas.drawCircle(x, y, radius, this.fillPaint);
     }
 
     @Override
-    public void fillPolygon(final float[] points) {
+    public void fillPolygon(final int[] points) {
         assert points != null;
         assert points.length >= 6;
         assert points.length % 2 == 0;
@@ -113,7 +113,7 @@ class AndroidCanvasImp implements CanvasImp {
     }
 
     @Override
-    public void fillRectangle(final float x, final float y, final float width, final float height) {
+    public void fillRectangle(final int x, final int y, final int width, final int height) {
         this.canvas.drawRect(x, y, x + width, y + height, this.fillPaint);
     }
 
@@ -240,7 +240,7 @@ class AndroidCanvasImp implements CanvasImp {
         return 0 <= x && x < this.getWidth() && 0 <= y && y < this.getHeight();
     }
 
-    private static Path createPath(final float[] points) {
+    private static Path createPath(final int[] points) {
         final Path result = new Path();
         boolean first = true;
         for (int i = 0; i < points.length; i = i + 2) {
